% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ecdf}
\alias{compute_ecdf}
\title{Compute ECDF (empirical cumulative distribution function)}
\usage{
compute_ecdf(posterior)
}
\arguments{
\item{posterior}{numeric vector of posterior probabilities over the prior support}
}
\value{
numeric vector with empirical cumulative distribution function 
(cumulative sum of \code{posterior})
}
\description{
Compute ECDF (empirical cumulative distribution function)
}
