% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_normalization_constant}
\alias{compute_normalization_constant}
\title{Compute normalization constant}
\usage{
compute_normalization_constant(counts, n_start, n_end, f_product)
}
\arguments{
\item{counts}{integer vector of counts}

\item{n_start}{start of prior support range}

\item{n_end}{end of prior support range}

\item{f_product}{product of (1-\code{fractions})}
}
\value{
normalization constant to compute posterior density
}
\description{
Compute normalization constant
}
