% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_posterior_with_replacement}
\alias{compute_posterior_with_replacement}
\title{Compute posterior probability with replacement}
\usage{
compute_posterior_with_replacement(n, counts, f_product, denominator)
}
\arguments{
\item{n}{integer for which to compute the posterior}

\item{counts}{integer vector of counts}

\item{f_product}{product of (1-\code{fractions})}

\item{denominator}{normalization constant returned by \code{compute_normalization_constant}}
}
\value{
posterior probability of \code{n}
}
\description{
Compute posterior probability with replacement
}
\seealso{
\link{compute_normalization_constant}
}
