% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_term}
\alias{compute_term}
\title{Compute single term (function F, Comoglio et al.)}
\usage{
compute_term(counts, n, f_product, t)
}
\arguments{
\item{counts}{integer vector of counts}

\item{n}{number of objects}

\item{f_product}{product of (1-\code{fractions})}

\item{t}{index vector}
}
\value{
single term of function F
}
\description{
Compute single term (function F, Comoglio et al.)
}
