% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Counts-class}
\alias{Counts-class}
\title{An S4 class to represent measurements (count data, sampling fractions)
and inferred population sizes}
\description{
An S4 class to represent measurements (count data, sampling fractions)
and inferred population sizes
}
\section{Slots}{

\describe{
\item{\code{counts}}{integer vector of counts (required)}

\item{\code{fractions}}{numeric vector of sampling fractions (required)}

\item{\code{n_start}}{start of prior support range. If omitted and total `counts` greater than zero,
computed as 0.5 * `mle`, where `mle` is the maximum likelihood estimate of the population size}

\item{\code{n_end}}{end of prior support range. If omitted and total `counts` greater than zero,
computed as 2 * `mle`, where `mle` is the maximum likelihood estimate of the population size}

\item{\code{f_product}}{product of (1-`fractions`)}

\item{\code{mle}}{maximum likelihood estimate of the population size (ratio between total counts and total sampling fraction)}

\item{\code{norm_constant}}{normalization constant}

\item{\code{posterior}}{numeric vector of posterior probabilities over the prior support}

\item{\code{map_p}}{maximum of `posterior` probability}

\item{\code{map_index}}{index of prior support corresponding to the maximum a posteriori}

\item{\code{map}}{maximum a posteriori of population size}

\item{\code{q_low}}{lower bound of the credible interval}

\item{\code{q_low_p}}{probability of the lower bound of the credible interval}

\item{\code{q_low_index}}{index of the prior support corresponding to `q_low`}

\item{\code{q_low_cum_p}}{cumulative posterior probability from `n1` to `q_low` (left tail)}

\item{\code{q_up}}{upper bound of the credible interval}

\item{\code{q_up_p}}{probability of the upper bound of the credible interval}

\item{\code{q_up_index}}{index of the prior support corresponding to `q_high`}

\item{\code{q_up_cum_p}}{cumulative posterior probability from `q_high` to `n2` (right tail)}

\item{\code{gamma}}{logical, TRUE if posterior computed using a Gamma approximation}
}}

\note{
The `posterior` slot contains either the PMF or a logical value used to 
compute posterior parameters with a Gamma approximation (see reference for details)

Lower and upper bounds of the credibile interval are computed at a default
confidence level of 95%

For more details on the normalization constant, see Corollary 1 in reference
}
\examples{
constructor:
# create an object of class 'Counts'
new_counts(counts = c(30, 35), fractions = c(0.075, 0.1))

# same, using `new`
new("Counts", counts = c(30, 35), fractions = c(0.075, 0.1))

}
\references{
Comoglio F, Fracchia L and Rinaldi M (2013) 
Bayesian inference from count data using discrete uniform priors. 
[PLoS ONE 8(10): e74388](http://www.plosone.org/article/info%3Adoi%2F10.1371%2Fjournal.pone.0074388)
}
\seealso{
[dupiR::new_counts()]
}
\author{
Federico Comoglio
}
\keyword{class}
