% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupiR-package.R
\docType{package}
\name{dupiR-package}
\alias{dupiR-package}
\alias{dupiR}
\title{Bayesian inference from count data using discrete uniform priors}
\description{
This package allows to infer population sizes using a binomial likelihood 
and least informative discrete uniform priors.
}
\references{
Comoglio F, Fracchia L, Rinaldi M (2013) Bayesian Inference from Count Data 
Using Discrete Uniform Priors. PLoS ONE 8(10): e74388
}
\author{
Federico Comoglio \email{federico.comoglio@gmail.com}

Maurizio Rinaldi
}
\keyword{package}
