% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_counts}
\alias{get_counts}
\title{Get \code{counts} slot for an object of class \code{Counts}}
\usage{
get_counts(object)
}
\arguments{
\item{object}{object of class \code{Counts}}
}
\value{
counts vector from a \code{Counts} object
}
\description{
Get \code{counts} slot for an object of class \code{Counts}
}
