% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_fractions}
\alias{get_fractions}
\title{Get \code{fractions} slot for an object of class \code{Counts}}
\usage{
get_fractions(object)
}
\arguments{
\item{object}{object of class \code{Counts}}
}
\value{
fractions vector from a \code{Counts} object
}
\description{
Get \code{fractions} slot for an object of class \code{Counts}
}
