% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_counts<-}
\alias{set_counts<-}
\title{Set \code{counts} slot for an object of class \code{Counts}}
\usage{
set_counts(object) <- value
}
\arguments{
\item{object}{object of class \code{Counts}}

\item{value}{numeric vector of counts}
}
\value{
an object of class \code{Counts}
}
\description{
Set \code{counts} slot for an object of class \code{Counts}
}
