% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_fractions<-}
\alias{set_fractions<-}
\title{Set \code{fractions} slot for an object of class \code{Counts}}
\usage{
set_fractions(object) <- value
}
\arguments{
\item{object}{object of class \code{Counts}}

\item{value}{numeric vector of sampling fractions}
}
\value{
an object of class \code{Counts}
}
\description{
Set \code{fractions} slot for an object of class \code{Counts}
}
