% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,Counts-method}
\alias{summary,Counts-method}
\title{Summary method for \code{Counts} class}
\usage{
\S4method{summary}{Counts}(object, ...)
}
\arguments{
\item{object}{object of class \code{Counts}}

\item{...}{additional parameters affecting the summary produced}
}
\value{
no return value, called for side effects
}
\description{
Summary method for \code{Counts} class
}
