% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_sum}
\alias{compute_sum}
\title{Compute sum of terms (function F, Comoglio et al.)}
\usage{
compute_sum(counts, n, f_product)
}
\arguments{
\item{counts}{integer vector of counts}

\item{n}{number of objects}

\item{f_product}{product of (1-\code{fractions})}
}
\value{
sum of terms in function F
}
\description{
Compute sum of terms (function F, Comoglio et al.)
}
