\name{dupiR-package}
\alias{dupiR-package}
\alias{dupiR}
\docType{package}
\title{Bayesian inference using discrete uniform priors with R
}
\description{Bayesian inference using discrete uniform priors with R
}
\details{
\tabular{ll}{
Package: \tab dupiR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-26\cr
License: \tab GPL-2\cr
}
}
\author{Federico Comoglio and Maurizio Rinaldi

Department of Biosystems Science and Engineering (D-BSSE)

ETH Zurich, Switzerland

and

Universita' del Piemonte Orientale 'A. Avogadro'

Maintainer: Federico Comoglio

\email{federico.comoglio@bsse.ethz.ch}	 
}

\keyword{package}
