\name{dupiR-package}
\alias{dupiR-package}
\alias{dupiR}
\docType{package}
\title{Bayesian inference using discrete uniform priors with R}
\description{This package implements a Bayesian approach to infer population sizes from count data. The package takes a set of sample counts obtained by sampling fractions of a finite volume containing an homogeneously dispersed population of identical objects and returns the posterior probability distribution of the population size. The algorithm makes use of a binomial likelihood and non-conjugate, discrete uniform priors. dupiR can be applied to both sampling with or without replacement.
}
\details{
\tabular{ll}{
Package: \tab dupiR\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-08-06\cr
License: \tab GPL-2\cr
}
}
\author{Federico Comoglio (1) and Maurizio Rinaldi (2)

(1) Department of Biosystems Science and Engineering (D-BSSE), ETH Zurich, Switzerland

(2) Dipartimento di Scienze del Farmaco, Universita' del Piemonte Orientale 'A. Avogadro'

Maintainer: Federico Comoglio

\email{federico.comoglio@bsse.ethz.ch}	 
}

\references{
Comoglio F, Fracchia L and Rinaldi M (2013) Bayesian inference from count data using discrete uniform priors. PLOS ONE, to appear
}
\keyword{package}