% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{initialize,Counts-method}
\alias{initialize,Counts-method}
\title{Initialize \code{Counts} class}
\usage{
\S4method{initialize}{Counts}(.Object, counts, fractions)
}
\arguments{
\item{.Object}{an object of class "Counts"}

\item{counts}{integer vector of counts}

\item{fractions}{numeric vector of sampling fractions}
}
\description{
Initialize \code{Counts} class
}
