% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,Counts-method}
\alias{plot,Counts-method}
\title{Plot method for \code{Counts} class}
\usage{
\S4method{plot}{Counts}(x, y, ...)
}
\arguments{
\item{x}{object of class \code{Counts}}

\item{y}{none}

\item{...}{additional parameters to be passed to \link{plot_posterior}}
}
\value{
no return value, called for side effects
}
\description{
Plot method for \code{Counts} class
}
