% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,Counts-method}
\alias{show,Counts-method}
\title{Print method for \code{Counts} class}
\usage{
\S4method{show}{Counts}(object)
}
\arguments{
\item{object}{object of class \code{Counts}}
}
\value{
no return value, called for side effects
}
\description{
Print method for \code{Counts} class
}
